If totp is null or empty in AuthenticateAsynd(), do not add the one_time_password element to the login payload.
Adjusted the sequence of loginResponse checks to consider "2fa_required" before looking for the "token".

This lets you attempt an authentication with a valid login/password but no totp.  This triggers the system to email if you do not have authenticator 2FA enabled.  You can then get the code they send you and try again providing that totp.

If you do have authenticator 2FA enabled, use the secret key to generate a totp and pass that in to authenticate in one pass.